+++ src/components/Drawer.js
@@ -1,19 +1,9 @@
-import React, { useState } from "react";
-
 import Drawer from "@mui/material/Drawer";
-// import MyModal from "./Modal";
 import List from "@mui/material/List";
-import { styled } from "@mui/material/styles";
+
 import ListItem from "@mui/material/ListItem";
 import Button from "@mui/material/Button";
 import ProviderInfo from "./ProviderInfo";
-import Modal from "@mui/material/Modal";
-import Box from "@mui/material/Box";
-import TableRow from "@mui/material/TableRow";
-import TableCell, { tableCellClasses } from "@mui/material/TableCell";
-import Divider from "@mui/material/Divider";
-import TableBody from "@mui/material/TableBody";
-import { Typography } from "@mui/material";
 
 const drawerWidth = -240;
 
@@ -45,56 +35,7 @@ const drawerWidth = -240;
 //   justifyContent: "flex-end",
 // }));
 
-export default function PersistentDrawerBottom({ comparedItems }) {
-  const [showModal, setShowModal] = useState(false);
-
-  const handleOpen = () => {
-    setShowModal(true);
-  };
-  const handleClose = () => {
-    setShowModal(false);
-  };
-  const style = {
-    position: "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    minWidth: 400,
-    maxWidth: "100%",
-    bgcolor: "background.paper",
-    border: "2px solid #000",
-    boxShadow: 24,
-    p: 4,
-  };
-  const StyledTableBody = styled(TableBody)(({ theme }) => ({
-    "&": {
-      borderRight: "1px solid purple",
-    },
-    // hide last border
-    "&:last-child , &:last-child ": {
-      borderRight: "none",
-    },
-  }));
-  const StyledTableRow = styled(TableRow)(({ theme }) => ({
-    "&:nth-of-type(odd)": {
-      backgroundColor: theme.palette.action.hover,
-    },
-    // hide last border
-    "&:last-child td, &:last-child th": {
-      border: 0,
-    },
-  }));
-  const StyledTableCell = styled(TableCell)(({ theme }) => ({
-    [`&.${tableCellClasses.head}`]: {
-      backgroundColor: theme.palette.common.black,
-      color: theme.palette.common.white,
-    },
-    [`&.${tableCellClasses.body}`]: {
-      fontSize: 14,
-      padding: "8px 48px",
-    },
-  }));
-
+export default function PersistentDrawerBottom({ comparedItems, handleOpen }) {
   return (
     <Drawer
       sx={{
@@ -112,7 +53,6 @@ export default function PersistentDrawerBottom({ comparedItems }) {
       open={true}
     >
       <List sx={{ display: "flex" }}>
-        {console.log("in drawer", comparedItems)}
         {comparedItems.map((el, index) => (
           <ListItem key={index}>
             <ProviderInfo
@@ -132,68 +72,6 @@ export default function PersistentDrawerBottom({ comparedItems }) {
           <></>
         )}
       </List>
-
-      <Modal
-        // hideBackdrop
-        open={showModal}
-        onClose={handleClose}
-        aria-labelledby="child-modal-title"
-        aria-describedby="child-modal-description"
-        comparedItems={comparedItems}
-      >
-        <Box sx={style}>
-          <Box>
-            <Typography component="h1" variant="h5" mb={2}>
-              Compare
-            </Typography>
-            <Divider />
-            <Box display="flex" sx={{ overflow: "auto" }}>
-              {comparedItems.map((item) => {
-                const miao = [
-                  {
-                    detail: `£ ${item.monthly_price}`,
-                    detailDescription: "Monthly Cost",
-                  },
-                  {
-                    detail: `${item.internet_speed} Mbps`,
-                    detailDescription: `${item.broadband_type} Speed`,
-                  },
-                  {
-                    detail: `£ ${item.set_up_cost}`,
-                    detailDescription: "Setup Costs",
-                  },
-                  {
-                    detail: `${item.contract_info}`,
-                    detailDescription: "Term End",
-                  },
-                  {
-                    detail: `${item.data_limits}`,
-                    detailDescription: "Data Limits",
-                  },
-                ];
-
-                return (
-                  <StyledTableBody>
-                    <Box p={2}>
-                      <ProviderInfo
-                        provider_name={item.provider_name}
-                        provider_logo_image_url={item.provider_logo_image_url}
-                        deal_name={item.deal_name}
-                      />
-                    </Box>
-                    {miao.map((x) => (
-                      <StyledTableRow>
-                        <StyledTableCell>{x.detailDescription}</StyledTableCell>
-                        <StyledTableCell>{x.detail}</StyledTableCell>
-                      </StyledTableRow>
-                    ))}
-                  </StyledTableBody>
-                );
-              })}
-            </Box>
-          </Box>
-        </Box>
-      </Modal>
     </Drawer>
   );
 }
